Use djxl's decoding loop in benchmark and tests.
This caught a bug in the rendering pipeline where pixel callbacks were not
reset between frames and the patch-frame of the second displayed frame was
still writing to the pixel callback set on the first frame.

Added preview frame handling to the ppf->codec io conversion function.

The reference one-shot decoding loop is in lib/extras/dec/jxl.cc now and it is
used in both djxl and the benchmark, lib/jxl/dec_file.cc was removed and
DecodeFile was added to test_utils.h and uses the same decoding function as
djxl.

A custom API-free DecodeFile had to be added back to render pipeline test, since
we don't want to expose the internal slow rendering pipeline through the API.
