+++ lib/extras/packed_image_convert.cc
@@ -20,6 +20,61 @@
 namespace jxl {
 namespace extras {
 
+Status ConvertPackedFrameToImageBundle(const JxlBasicInfo& info,
+                                       const PackedFrame& frame,
+                                       const CodecInOut& io, ThreadPool* pool,
+                                       ImageBundle* bundle) {
+  JXL_ASSERT(frame.color.pixels() != nullptr);
+  size_t frame_bits_per_sample =
+      (frame.color.bitdepth_from_format
+           ? frame.color.BitsPerChannel(frame.color.format.data_type)
+           : info.bits_per_sample);
+  JXL_ASSERT(frame_bits_per_sample != 0);
+  // It is ok for the frame.color.format.num_channels to not match the
+  // number of channels on the image.
+  JXL_ASSERT(1 <= frame.color.format.num_channels &&
+             frame.color.format.num_channels <= 4);
+
+  const Span<const uint8_t> span(
+      static_cast<const uint8_t*>(frame.color.pixels()),
+      frame.color.pixels_size);
+  Rect frame_rect = Rect(
+      frame.frame_info.layer_info.crop_x0, frame.frame_info.layer_info.crop_y0,
+      frame.frame_info.layer_info.xsize, frame.frame_info.layer_info.ysize);
+  JXL_ASSERT(frame_rect.IsInside(Rect(0, 0, info.xsize, info.ysize)));
+  if (info.have_animation) {
+    bundle->duration = frame.frame_info.duration;
+    bundle->blend = frame.frame_info.layer_info.blend_info.blendmode > 0;
+    bundle->use_for_next_frame =
+        frame.frame_info.layer_info.save_as_reference > 0;
+    bundle->origin.x0 = frame.frame_info.layer_info.crop_x0;
+    bundle->origin.y0 = frame.frame_info.layer_info.crop_y0;
+  }
+  bundle->name = frame.name;  // frame.frame_info.name_length is ignored here.
+  JXL_ASSERT(io.metadata.m.color_encoding.IsGray() ==
+             (frame.color.format.num_channels <= 2));
+
+  const bool float_in = frame.color.format.data_type == JXL_TYPE_FLOAT16 ||
+                        frame.color.format.data_type == JXL_TYPE_FLOAT;
+  JXL_RETURN_IF_ERROR(ConvertFromExternal(
+      span, frame.color.xsize, frame.color.ysize, io.metadata.m.color_encoding,
+      frame.color.format.num_channels,
+      /*alpha_is_premultiplied=*/info.alpha_premultiplied,
+      frame_bits_per_sample, frame.color.format.endianness,
+      /*flipped_y=*/frame.color.flipped_y, pool, bundle,
+      /*float_in=*/float_in, /*align=*/0));
+
+  bundle->extra_channels().resize(io.metadata.m.extra_channel_info.size());
+  for (size_t i = 0; i < frame.extra_channels.size(); i++) {
+    const auto& ppf_ec = frame.extra_channels[i];
+    bundle->extra_channels()[i] = ImageF(ppf_ec.xsize, ppf_ec.ysize);
+    JXL_CHECK(BufferToImageF(ppf_ec.format, ppf_ec.xsize, ppf_ec.ysize,
+                             ppf_ec.pixels(), ppf_ec.pixels_size, pool,
+                             &bundle->extra_channels()[i]));
+  }
+  return true;
+}
+
 Status ConvertPackedPixelFileToCodecInOut(const PackedPixelFile& ppf,
                                           ThreadPool* pool, CodecInOut* io) {
   const bool has_alpha = ppf.info.alpha_bits != 0;
@@ -105,61 +160,24 @@ Status ConvertPackedPixelFileToCodecInOut(const PackedPixelFile& ppf,
     io->metadata.m.extra_channel_info.push_back(std::move(out));
   }
 
+  // Convert the preview
+  if (ppf.preview_frame.get()) {
+    size_t preview_xsize = ppf.preview_frame->color.xsize;
+    size_t preview_ysize = ppf.preview_frame->color.ysize;
+    io->metadata.m.have_preview = true;
+    JXL_RETURN_IF_ERROR(
+        io->metadata.m.preview_size.Set(preview_xsize, preview_ysize));
+    JXL_RETURN_IF_ERROR(ConvertPackedFrameToImageBundle(
+        ppf.info, *ppf.preview_frame, *io, pool, &io->preview_frame));
+  }
+
   // Convert the pixels
   io->dec_pixels = 0;
   io->frames.clear();
   for (const auto& frame : ppf.frames) {
-    JXL_ASSERT(frame.color.pixels() != nullptr);
-    size_t frame_bits_per_sample =
-        (frame.color.bitdepth_from_format
-             ? frame.color.BitsPerChannel(frame.color.format.data_type)
-             : ppf.info.bits_per_sample);
-    JXL_ASSERT(frame_bits_per_sample != 0);
-    // It is ok for the frame.color.format.num_channels to not match the
-    // number of channels on the image.
-    JXL_ASSERT(1 <= frame.color.format.num_channels &&
-               frame.color.format.num_channels <= 4);
-
-    const Span<const uint8_t> span(
-        static_cast<const uint8_t*>(frame.color.pixels()),
-        frame.color.pixels_size);
-    Rect frame_rect = Rect(frame.frame_info.layer_info.crop_x0,
-                           frame.frame_info.layer_info.crop_y0,
-                           frame.frame_info.layer_info.xsize,
-                           frame.frame_info.layer_info.ysize);
-    JXL_ASSERT(frame_rect.IsInside(Rect(0, 0, ppf.info.xsize, ppf.info.ysize)));
     ImageBundle bundle(&io->metadata.m);
-    if (ppf.info.have_animation) {
-      bundle.duration = frame.frame_info.duration;
-      bundle.blend = frame.frame_info.layer_info.blend_info.blendmode > 0;
-      bundle.use_for_next_frame =
-          frame.frame_info.layer_info.save_as_reference > 0;
-      bundle.origin.x0 = frame.frame_info.layer_info.crop_x0;
-      bundle.origin.y0 = frame.frame_info.layer_info.crop_y0;
-    }
-    bundle.name = frame.name;  // frame.frame_info.name_length is ignored here.
-    JXL_ASSERT(io->metadata.m.color_encoding.IsGray() ==
-               (frame.color.format.num_channels <= 2));
-
-    const bool float_in = frame.color.format.data_type == JXL_TYPE_FLOAT16 ||
-                          frame.color.format.data_type == JXL_TYPE_FLOAT;
-    JXL_RETURN_IF_ERROR(ConvertFromExternal(
-        span, frame.color.xsize, frame.color.ysize,
-        io->metadata.m.color_encoding, frame.color.format.num_channels,
-        /*alpha_is_premultiplied=*/ppf.info.alpha_premultiplied,
-        frame_bits_per_sample, frame.color.format.endianness,
-        /*flipped_y=*/frame.color.flipped_y, pool, &bundle,
-        /*float_in=*/float_in, /*align=*/0));
-
-    bundle.extra_channels().resize(io->metadata.m.extra_channel_info.size());
-    for (size_t i = 0; i < frame.extra_channels.size(); i++) {
-      const auto& ppf_ec = frame.extra_channels[i];
-      bundle.extra_channels()[i] = ImageF(ppf_ec.xsize, ppf_ec.ysize);
-      JXL_CHECK(BufferToImageF(ppf_ec.format, ppf_ec.xsize, ppf_ec.ysize,
-                               ppf_ec.pixels(), ppf_ec.pixels_size, pool,
-                               &bundle.extra_channels()[i]));
-    }
-
+    JXL_RETURN_IF_ERROR(
+        ConvertPackedFrameToImageBundle(ppf.info, frame, *io, pool, &bundle));
     io->frames.push_back(std::move(bundle));
     io->dec_pixels += frame.color.xsize * frame.color.ysize;
   }
