+++ lib/jxl/base/data_parallel.h
@@ -31,6 +31,9 @@ class ThreadPool {
   ThreadPool(const ThreadPool&) = delete;
   ThreadPool& operator&(const ThreadPool&) = delete;
 
+  JxlParallelRunner runner() const { return runner_; }
+  void* runner_opaque() const { return runner_opaque_; }
+
   // Runs init_func(num_threads) followed by data_func(task, thread) on worker
   // thread(s) for every task in [begin, end). init_func() must return a Status
   // indicating whether the initialization succeeded.
