+++ lib/jxl/speed_tier_test.cc
@@ -12,8 +12,6 @@
 #include "lib/jxl/base/padded_bytes.h"
 #include "lib/jxl/base/thread_pool_internal.h"
 #include "lib/jxl/codec_in_out.h"
-#include "lib/jxl/dec_file.h"
-#include "lib/jxl/dec_params.h"
 #include "lib/jxl/enc_butteraugli_comparator.h"
 #include "lib/jxl/enc_cache.h"
 #include "lib/jxl/enc_file.h"
@@ -98,10 +96,9 @@ TEST_P(SpeedTierTest, Roundtrip) {
 
   CompressParams cparams;
   cparams.speed_tier = params.speed_tier;
-  DecompressParams dparams;
 
   CodecInOut io2;
-  test::Roundtrip(&io, cparams, dparams, nullptr, &io2);
+  test::Roundtrip(&io, cparams, {}, nullptr, &io2);
 
   // Can be 2.2 in non-hare mode.
   EXPECT_LE(ButteraugliDistance(io, io2, cparams.ba_params, GetJxlCms(),
