+++ lib/jxl/splines_test.cc
@@ -7,7 +7,6 @@
 
 #include "lib/extras/codec.h"
 #include "lib/jxl/base/printf_macros.h"
-#include "lib/jxl/dec_file.h"
 #include "lib/jxl/enc_butteraugli_comparator.h"
 #include "lib/jxl/enc_color_management.h"
 #include "lib/jxl/enc_splines.h"
@@ -327,8 +326,8 @@ TEST(SplinesTest, ClearedEveryFrame) {
   CodecInOut io_actual;
   const PaddedBytes bytes_actual =
       ReadTestData("jxl/spline_on_first_frame.jxl");
-  ASSERT_TRUE(DecodeFile(DecompressParams(), bytes_actual, &io_actual,
-                         /*pool=*/nullptr));
+  ASSERT_TRUE(test::DecodeFile({}, bytes_actual, &io_actual,
+                               /*pool=*/nullptr));
 
   ASSERT_TRUE(io_actual.TransformTo(ColorEncoding::SRGB(), GetJxlCms()));
   for (size_t c = 0; c < 3; ++c) {
