+++ lib/lib.gni
@@ -108,7 +108,6 @@ libjxl_dec_sources = [
     "jxl/dec_modular.h",
     "jxl/dec_noise.cc",
     "jxl/dec_noise.h",
-    "jxl/dec_params.h",
     "jxl/dec_patch_dictionary.cc",
     "jxl/dec_patch_dictionary.h",
     "jxl/dec_tone_mapping-inl.h",
@@ -258,8 +257,6 @@ libjxl_enc_sources = [
     "jxl/butteraugli/butteraugli.cc",
     "jxl/butteraugli/butteraugli.h",
     "jxl/butteraugli_wrapper.cc",
-    "jxl/dec_file.cc",
-    "jxl/dec_file.h",
     "jxl/enc_ac_strategy.cc",
     "jxl/enc_ac_strategy.h",
     "jxl/enc_adaptive_quantization.cc",
@@ -451,6 +448,8 @@ libjxl_extras_sources = [
     "extras/dec/color_hints.h",
     "extras/dec/decode.cc",
     "extras/dec/decode.h",
+    "extras/dec/jxl.cc",
+    "extras/dec/jxl.h",
     "extras/dec/pgx.cc",
     "extras/dec/pgx.h",
     "extras/dec/pnm.cc",
