+++ tools/benchmark/benchmark_codec_jxl.cc
@@ -12,9 +12,10 @@
 #include <utility>
 #include <vector>
 
-#include "jxl/decode_cxx.h"
 #include "jxl/thread_parallel_runner_cxx.h"
 #include "lib/extras/codec.h"
+#include "lib/extras/dec/jxl.h"
+#include "lib/extras/packed_image_convert.h"
 #include "lib/extras/time.h"
 #include "lib/jxl/aux_out.h"
 #include "lib/jxl/base/data_parallel.h"
@@ -22,8 +23,6 @@
 #include "lib/jxl/base/padded_bytes.h"
 #include "lib/jxl/base/span.h"
 #include "lib/jxl/codec_in_out.h"
-#include "lib/jxl/dec_file.h"
-#include "lib/jxl/dec_params.h"
 #include "lib/jxl/enc_cache.h"
 #include "lib/jxl/enc_color_management.h"
 #include "lib/jxl/enc_external_image.h"
@@ -274,121 +273,25 @@ class JxlCodec : public ImageCodec {
                     const Span<const uint8_t> compressed,
                     ThreadPoolInternal* pool, CodecInOut* io,
                     jpegxl::tools::SpeedStats* speed_stats) override {
-    io->frames.clear();
-    if (dparams_.max_passes != DecompressParams().max_passes ||
-        dparams_.max_downsampling != DecompressParams().max_downsampling) {
-      // Must use the C++ API to honor non-default dparams.
-      if (uint8_) {
-        return JXL_FAILURE(
-            "trying to use decompress params that are not all available in "
-            "either decoding API");
-      }
-      const double start = Now();
-      JXL_RETURN_IF_ERROR(DecodeFile(dparams_, compressed, io, pool));
-      const double end = Now();
-      speed_stats->NotifyElapsed(end - start);
-      return true;
-    }
-
-    double elapsed_convert_image = 0;
+    dparams_.runner = pool->runner();
+    dparams_.runner_opaque = pool->runner_opaque();
+    JxlDataType data_type = uint8_ ? JXL_TYPE_UINT8 : JXL_TYPE_FLOAT;
+    dparams_.accepted_formats = {{3, data_type, JXL_NATIVE_ENDIAN, 0},
+                                 {4, data_type, JXL_NATIVE_ENDIAN, 0}};
+    // By default, the decoder will undo exif orientation, giving an image
+    // with identity exif rotation as result. However, the benchmark does
+    // not undo exif orientation of the originals, and compares against the
+    // originals, so we must set the option to keep the original orientation
+    // instead.
+    dparams_.keep_orientation = true;
+    extras::PackedPixelFile ppf;
+    size_t decoded_bytes;
     const double start = Now();
-    {
-      std::vector<uint8_t> pixel_data;
-      PaddedBytes icc_profile;
-      auto runner = JxlThreadParallelRunnerMake(nullptr, pool->NumThreads());
-      auto dec = JxlDecoderMake(nullptr);
-      // By default, the decoder will undo exif orientation, giving an image
-      // with identity exif rotation as result. However, the benchmark does
-      // not undo exif orientation of the originals, and compares against the
-      // originals, so we must set the option to keep the original orientation
-      // instead.
-      JxlDecoderSetKeepOrientation(dec.get(), JXL_TRUE);
-      JXL_RETURN_IF_ERROR(
-          JXL_DEC_SUCCESS ==
-          JxlDecoderSubscribeEvents(dec.get(), JXL_DEC_BASIC_INFO |
-                                                   JXL_DEC_COLOR_ENCODING |
-                                                   JXL_DEC_FULL_IMAGE));
-      JxlBasicInfo info{};
-      JxlPixelFormat format = {/*num_channels=*/3,
-                               /*data_type=*/JXL_TYPE_FLOAT,
-                               /*endianness=*/JXL_NATIVE_ENDIAN,
-                               /*align=*/0};
-      if (uint8_) {
-        format.data_type = JXL_TYPE_UINT8;
-      }
-      JxlDecoderSetInput(dec.get(), compressed.data(), compressed.size());
-      JxlDecoderStatus status;
-      while ((status = JxlDecoderProcessInput(dec.get())) != JXL_DEC_SUCCESS) {
-        switch (status) {
-          case JXL_DEC_ERROR:
-            return JXL_FAILURE("decoder error");
-          case JXL_DEC_NEED_MORE_INPUT:
-            return JXL_FAILURE("decoder requests more input");
-          case JXL_DEC_BASIC_INFO:
-            JXL_RETURN_IF_ERROR(JXL_DEC_SUCCESS ==
-                                JxlDecoderGetBasicInfo(dec.get(), &info));
-            format.num_channels = info.num_color_channels;
-            if (info.alpha_bits != 0) {
-              ++format.num_channels;
-              io->metadata.m.extra_channel_info.resize(1);
-              io->metadata.m.extra_channel_info[0].type =
-                  jxl::ExtraChannel::kAlpha;
-            }
-            break;
-          case JXL_DEC_COLOR_ENCODING: {
-            size_t icc_size;
-            JXL_RETURN_IF_ERROR(JXL_DEC_SUCCESS ==
-                                JxlDecoderGetICCProfileSize(
-                                    dec.get(), &format,
-                                    JXL_COLOR_PROFILE_TARGET_DATA, &icc_size));
-            icc_profile.resize(icc_size);
-            JXL_RETURN_IF_ERROR(JXL_DEC_SUCCESS ==
-                                JxlDecoderGetColorAsICCProfile(
-                                    dec.get(), &format,
-                                    JXL_COLOR_PROFILE_TARGET_DATA,
-                                    icc_profile.data(), icc_profile.size()));
-            break;
-          }
-          case JXL_DEC_NEED_IMAGE_OUT_BUFFER: {
-            size_t buffer_size;
-            JXL_RETURN_IF_ERROR(
-                JXL_DEC_SUCCESS ==
-                JxlDecoderImageOutBufferSize(dec.get(), &format, &buffer_size));
-            JXL_RETURN_IF_ERROR(buffer_size ==
-                                info.xsize * info.ysize * format.num_channels *
-                                    (uint8_ ? sizeof(uint8_t) : sizeof(float)));
-            pixel_data.resize(buffer_size);
-            JXL_RETURN_IF_ERROR(JXL_DEC_SUCCESS ==
-                                JxlDecoderSetImageOutBuffer(dec.get(), &format,
-                                                            pixel_data.data(),
-                                                            buffer_size));
-            break;
-          }
-          case JXL_DEC_FULL_IMAGE: {
-            const double start_convert_image = Now();
-            {
-              ColorEncoding color_encoding;
-              JXL_RETURN_IF_ERROR(
-                  color_encoding.SetICC(PaddedBytes(icc_profile)));
-              ImageBundle frame(&io->metadata.m);
-              JXL_RETURN_IF_ERROR(BufferToImageBundle(
-                  format, info.xsize, info.ysize, pixel_data.data(),
-                  pixel_data.size(), pool, color_encoding, &frame));
-              io->frames.push_back(std::move(frame));
-              io->dec_pixels += info.xsize * info.ysize;
-            }
-            const double end_convert_image = Now();
-            elapsed_convert_image += end_convert_image - start_convert_image;
-            break;
-          }
-          default:
-            return JXL_FAILURE("unrecognized status %d",
-                               static_cast<int>(status));
-        }
-      }
-    }
+    JXL_RETURN_IF_ERROR(DecodeImageJXL(compressed.data(), compressed.size(),
+                                       dparams_, &decoded_bytes, &ppf));
     const double end = Now();
-    speed_stats->NotifyElapsed(end - start - elapsed_convert_image);
+    speed_stats->NotifyElapsed(end - start);
+    JXL_RETURN_IF_ERROR(ConvertPackedPixelFileToCodecInOut(ppf, pool, io));
     return true;
   }
 
@@ -403,7 +306,7 @@ class JxlCodec : public ImageCodec {
   AuxOut cinfo_;
   CompressParams cparams_;
   bool has_ctransform_ = false;
-  DecompressParams dparams_;
+  extras::JXLDecompressParams dparams_;
   bool uint8_ = false;
 };
